Move timeline view to own file.
